+++ Projects/KidsCare_UI/src/main/java/kr/ac/kopo/kidscare/model/KCUserPost.java
@@ -23,7 +23,7 @@ public class KCUserPost {
 	}
 	
 
-	public Long getUserPostId() {
+	public Integer getUserPostId() {
 
 		return userPostId;
 	}
